v0.1.7.1 - huge rebalance, move mechnanics changed
 - now hero ship cannot change move direction instantly
 - all structures can now be build on all spots(but priorities for inner/outer circles saved)
 - dome now costs 500 gold
 - gold income decreased(0.0055->0.004)
 - research income increased(0.0001->0.0005)
 - hero view scale changed to 2
 - new skill - "shot" for 1 gold

+++ JustTest/UnitInfo.h
@@ -221,7 +221,7 @@ class UnitInfo {
 			}
 			
 			if (setting == "default_speed") {
-				default_speed = std::stoi(value); 
+				default_speed = std::stof(value); 
 				if (settings.isUnitInfoDebugEnabled()) {
 					std::cout << "Default speed is set to " << default_speed << std::endl;
 				}
